requirements.txt
Django==4.2.13
django-crispy-forms==2.0
crispy-bootstrap4==2022.1
Pillow==10.3.0
requests==2.32.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
django-widget-tweaks==1.4.12
pymysql==1.1.1
cryptography==43.0.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
